refactor: column, columnOptions test
